Making necessary changes and bug fixes to CC32 code.
Updated name of the UserClient class so that ORCA
will correctly initialize it.  Fixed initialization
codes:

0x4186 -> 0x4186

to correspond to manual.  

Fixed generation of Q and F codes to grab the slot
number of the CC32 as specified by the manual.  All
of this has been tested.  

Known issues:
NAF commands directed towards the CC32 (commands 
with N=0 or N>=24) are passed to the CC32 but they
do *NOT* set the Q and X bits.  For all NAF commands
ORCA checks the status of the X bit and throws an
exception if the X bit is not set after a command.  

I think this is because the X bit should be cleared 
(i.e. reset) after any failed bus command and also
it should be set in the initialization code of the CC32
before any command directed towards the CC32 is initiated
so that ORCA doesn't think the commands have failed
when they have in fact succeeded.



git-svn-id: svn://orca.physics.unc.edu/Orca@1550 fd86f90f-2c0a-422a-b972-84106d8db31b