nEDMCoil Updates
Committing a working copy of the nEDM Coil Process.  This runs, further testing is underway.

Updating PID control of nEDM coils

Adding realProcessTime to track the actual speed of the cycle.  Adding to UI.

Formatting changes.

In _checkAndClearErrors, first call the non-blocking check and *then* go back and wait for all commands to be done.  This is *much* more efficienct.

Using non-blocking readback, since the code waits right after this function.

Calculating the correct delay to call the _runProcess function again.  This keeps the function running very close to the input frequency.

Ensuring the we clear the CurrentMemory after a run.

Adding check for MagnetometerMap count

Adding readback and resetTrips for all power supplies.

Allowing targetField to be set externally

Enabling loading and setting target field via plist file.  Also enable saving current field to a plist file.

Limiting files that may be read in in _readFile, and adding _saveFile.

Adding updates to the GUI interface to take advantage of the target field API.

Bug in nEDMExperiment, not incrementing index

git-svn-id: svn://orca.physics.unc.edu/Orca@8967 fd86f90f-2c0a-422a-b972-84106d8db31b