Add counter advanced options & bind forced update
This commit adds a box that contains the options that
enable/disable the following counter options.
Lead Zero Blanking
Test Mode
Inhibit Counting

I also bound the GUI button for choosing to force updates
to a model parameter.
I plan on implementing the functionality for this in the
next commit (or two).