SNO+: update from SNO+ settings when object is added.
Previously, when the SNO+ settings tab was updated, the SNO+ model would update
all the relevant models (MTC, CAEN, and XL3), but if an object was later added
the settings would not be in sync.

Now, the MTC, CAEN, and XL3 objects look for a notification to be posted when
objects are added, and if it is themselves, then they update these settings
from the SNO+ model.

git-svn-id: svn://orca.physics.unc.edu/Orca@10537 fd86f90f-2c0a-422a-b972-84106d8db31b