SNO+: update how global settings are synced
Previously the models subscribed to a notification sent out whenever
an object is added and checked to see if it was them, if so update settings
from SNO+ model. This was inefficient, and was getting called also when
ORCA started (in which case the settings are loaded after the document
is loaded). Now, the models just try to update in initWithCoder() *and*
in awakeAfterDocumentLoaded() which should guarantee that they are synced
whether ORCA just started up or they were just added to the experiment.

git-svn-id: svn://orca.physics.unc.edu/Orca@10540 fd86f90f-2c0a-422a-b972-84106d8db31b