SNO+: fix bug when standard run type or version is NULL
Fixes the crash when the standard run type or standard run version fields
retrieved from ORCA DB are NULL. It sets a timeout of 2 seconds for Couch DB
to be reached when updating the standard run. It also minimizes the print
out statements and cleans up the code.

Pull request from Javi Caravaca.

git-svn-id: svn://orca.physics.unc.edu/Orca@10560 fd86f90f-2c0a-422a-b972-84106d8db31b