SNO+: update SNO+ model to get run number from database
When a new run starts, the SNO+ model will now get the next run from the
postgres database. If the database is not available, it will abort the run. If
the query fails for some reason, the run number will be set to 0xffffffff,
which will be interpreted by the builder as a default run and saved to
default_[something].zdab. This means we will still see events in XSNOED, but
the runs will not be processed by the data flow scripts or the L2 trigger.

Also fixes a few bugs in ORPQModel, and adds new methods for querying the
database without a timeout and without a selector.

Some minor fixes to ORRunModel where the run number was being interpreted as a
signed integer instead of an unsigned integer.

Pull request from Anthony LaTorre (with some code from Phil Harvey).

git-svn-id: svn://orca.physics.unc.edu/Orca@10866 fd86f90f-2c0a-422a-b972-84106d8db31b