Fix bug in setting a run time limit.
Previously the run time limit field might interpret a number with a
comma in it (e.g. 9,999) as a string and then convert the comma to an
period (9,999->9.999) meaning you couldn't have times in excess of 1000
seconds.

This commit fixes that by change the run time limit NSTextField to a
NSTextField with NumberFormatter. The NumberFormatter is not customized
in any way.