SNO+: Fix user-lockout of HV control
The methods to read back voltage/current and switch status can
return without throwing an exception in certain cases. If reading
switch status failed in this way but then reading back voltage was
successful, the hv control logic would end up in a state where the
monitoring thread was running but the user was unable to control
the high voltage. This ensures that the switch and voltage are both
successfully read back before starting the high voltage thread,
preventing the above from occuring.

Pull request from Ben Land.

git-svn-id: svn://orca.physics.unc.edu/Orca@10896 fd86f90f-2c0a-422a-b972-84106d8db31b