SNO+: Add HV parameter readback from postgres.
Asks the user once to add the postgres object to the experiment if
it does not eixst. Asks per crate to use hardcoded defaults if an
error is encountered with the SQL query. Potentially asks multiple
times if the user declines hardcoded values and the query continues
to fail.

Pull request by Ben Land