SNO+: update XL3 model to wait until crate has been initialized before checking the HV
The HV registers are undefined just after the crate has been turned on but
before the crate has been reset. So, the HV thread now waits until the crate
has been reset before continuing. This commit also updates the crate reset
button so that it is enabled only when either the crate has not been reset or
if the HV is off.

Author: Anthony LaTorre <tlatorre@uchicago.edu>

git-svn-id: svn://orca.physics.unc.edu/Orca@10918 fd86f90f-2c0a-422a-b972-84106d8db31b