Delete unused status readout items
Since data no longer flows through orca these values are not easily
available any longer. So there's no reason they should remain in code.

Note this does however break the ORVME readout for the MTC. This
does not affect SNO+ now that the MTC server handles the VME interface.
But in the unlikely event that we want ORCA to handle the VME stuff
again this will be a problem