SNO+: update HV target field if necessary.
If the readback on start is higher than the saved target value also
update the target value (and the GUI as a consequence) to the
readback to ensure that the target is always >= the setpoint. This
fixes a popup message informing the user that the target must be
greater than the setpoint upon first switching to the XL3 window
after opening a .orca file with a target saved below the readback.