SNO+: fix a bug in CAEN initialization
This commit fixes a bug in the initBoard function of the CAEN model.
Previously, the acquisition control register was written before performing a
software reset, which causes all of the registers to be set to their default
values, thus overwriting the settings in the acquisition control register. Now,
the software reset is performed first and the acquisition control register
written afterwards.

Author: Anthony LaTorre <tlatorre@uchicago.edu>