SNO+: don't send a queue_run_start command for rollover runs
This commit updates the runInitialization method in the SNO+ model so that it
doesn't send the mtc server a queue_run_start command. This will cause the mtc
server to set the first and valid gtid fields in the run header to be the same,
which is how analysis will know that it was a rollover run.