SNO+: Read files describing standard TELLIE flash sequences.
This commit updates the SNOPController such that documents
describing standard TELLIE flash sequences can be read from couchdb
and run with a single button click. Additonally, the ELLIE gui has
been updated so such documents can be custom generated and pushed
to the database. Cosmetic updates also made to the SMELLIE tab in
the SNOPController.