SNO+:Add next/prev hot-keys to several views
Any views that have arrow buttons to swap to next/previous item.
In general I use vim-like 'h' or 'l' for left, right. B/c the
left/right arrow keys were already taken. But for a few view where
they're not taken I used arrow keys as well.

The affected views are as follows
- SNOCrate uses 'h' and 'l'
- SNORack uses 'h' and 'l'
- Xl3 uses left, right, 'h', and 'l'
- FEC uses 'h' and 'l'
- DB uses left, right, 'h', and 'l'

Also adds a way to 'esc' out of a text input field for the MTC, XL3 and
DB gui

Author: Eric Marzec <marzece@sas.upenn.edu>

git-svn-id: svn://orca.physics.unc.edu/Orca@11089 fd86f90f-2c0a-422a-b972-84106d8db31b