SNO+: Fix inconsistency in units used on TUBii.
Previously, TUBii used ns for delay lengths and s for pulse widths. This was
confusing and inconsistent. This pull request accompanies a firmware update
which changes the pulse widths to use ns. The gui is changed purely to display
the correct units while the ELLIE model, which actually uses these pulsers, has
its commands updated.

Author: Ian Coulter <icoulter@hep.upenn.edu>

git-svn-id: svn://orca.physics.unc.edu/Orca@11094 fd86f90f-2c0a-422a-b972-84106d8db31b