SNO+: update SNO+ model and run model to load elapsed time since run start
This commit updates the SNO+ model so that it loads the elapsed time since the
run started from the detector state database to the run model on boot up. This
required a small change to the run control to start the run timer so that the
elapsed time is incremented properly.