SNO+: Orca version checking at startup
When Orca starts for SNO+, check the current version (in SNOPModel.h)
against the current production version, stored in the Orca session
Postgres DB.

The interface to this database (which now handles version checking,
exclusive lock checking, and posting session information) is now
encapsulated within a new class `SessionDB` which the `SNOPModel`
uses as a helper, to help keep the implementation in the `SNOPModel`
under control.

Also, change 'hostile takeover' button wording to clarify.

git-svn-id: svn://orca.physics.unc.edu/Orca@11429 fd86f90f-2c0a-422a-b972-84106d8db31b